M10_ReAct_Pattern/requirements.txt
# Module 10.1: ReAct Pattern Implementation
# Core dependencies for agentic RAG with ReAct pattern

# LangChain for agent orchestration
langchain==0.1.20
langchain-openai==0.0.8
langchain-community==0.0.38

# OpenAI for LLM
openai==1.12.0

# Vector database (from Level 1)
pinecone-client==3.0.3

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Monitoring (optional)
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
